Issue #224 - Load lib classes before lib vars
Preload lib classes prior to lib vars so that vars can use lib classes
as method argument types without spurious type mismatches and
`MissingMethodException`s. This was an odd side effect of the
class preload logic in 1.4's 5c90975 and 6f2639f, which caused the
classes to be loaded twice, confusing the method interceptor's
matching logic.